pyproject.toml
[tool.poetry]
name = "load-test"
version = "0.1.0"
description = ""
authors = ["Victor Kayk <victorkayk77@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
faker = "^19.13.0"
locust = "^2.18.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
